+++ src/hal.rs
@@ -29,8 +29,7 @@
 use core::convert::Infallible;
 use std::time::{Duration, Instant};
 
-use embedded_hal::delay::blocking::{DelayMs, DelayUs};
-use embedded_hal::timer::nb::CountDown;
+use embedded_hal::delay::blocking::{DelayUs};
 use spin_sleep::sleep;
 use void::Void;
 
@@ -38,7 +37,7 @@ use void::Void;
 #[derive(Debug, Default)]
 pub struct Delay;
 
-/// `Delay` trait implementation for `embedded-hal` v1.0.0-alpha.5.
+/// `Delay` trait implementation for `embedded-hal` v1.0.0-alpha.7.
 impl Delay {
     /// Constructs a new `Delay`.
     pub fn new() -> Delay {
@@ -46,114 +45,64 @@ impl Delay {
     }
 }
 
-/// `DelayMs<u8>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayMs<u8> for Delay {
-    type Error = Infallible;
-
-    fn delay_ms(&mut self, ms: u8) -> Result<(), Self::Error> {
-        sleep(Duration::from_millis(u64::from(ms)));
-        Ok(())
-    }
-}
-
 /// `DelayMs<u8>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayMs<u8> for Delay {
     fn delay_ms(&mut self, ms: u8) {
-        DelayMs::delay_ms(self, ms).unwrap()
-    }
-}
-
-/// `DelayMs<u16>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayMs<u16> for Delay {
-    type Error = Infallible;
-
-    fn delay_ms(&mut self, ms: u16) -> Result<(), Self::Error> {
-        sleep(Duration::from_millis(u64::from(ms)));
-        Ok(())
+        DelayUs::delay_ms(self, ms as u32).unwrap()
     }
 }
 
 /// `DelayMs<u16>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayMs<u16> for Delay {
     fn delay_ms(&mut self, ms: u16) {
-        DelayMs::delay_ms(self, ms).unwrap()
-    }
-}
-
-/// `DelayMs<u32>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayMs<u32> for Delay {
-    type Error = Infallible;
-
-    fn delay_ms(&mut self, ms: u32) -> Result<(), Self::Error> {
-        sleep(Duration::from_millis(u64::from(ms)));
-        Ok(())
+        DelayUs::delay_ms(self, ms as u32).unwrap()
     }
 }
 
 /// `DelayMs<u32>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayMs<u32> for Delay {
     fn delay_ms(&mut self, ms: u32) {
-        DelayMs::delay_ms(self, ms).unwrap()
-    }
-}
-
-/// `DelayMs<u64>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayMs<u64> for Delay {
-    type Error = Infallible;
-
-    fn delay_ms(&mut self, ms: u64) -> Result<(), Self::Error> {
-        sleep(Duration::from_millis(ms));
-        Ok(())
+        DelayUs::delay_ms(self, ms).unwrap()
     }
 }
 
 /// `DelayMs<u64>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayMs<u64> for Delay {
-    fn delay_ms(&mut self, ms: u64) {
-        DelayMs::delay_ms(self, ms).unwrap()
-    }
-}
+    fn delay_ms(&mut self, mut ms: u64) {
+      while ms > (u32::MAX as u64) {
+          ms -= u32::MAX as u64;
+          DelayUs::delay_ms(self, u32::MAX).unwrap();
+      }
 
-/// `DelayUs<u8>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayUs<u8> for Delay {
-    type Error = Infallible;
-
-    fn delay_us(&mut self, us: u8) -> Result<(), Self::Error> {
-        sleep(Duration::from_micros(us.into()));
-        Ok(())
+      DelayUs::delay_ms(self, ms as u32).unwrap()
     }
 }
 
 /// `DelayUs<u8>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayUs<u8> for Delay {
     fn delay_us(&mut self, us: u8) {
-        DelayUs::delay_us(self, us).unwrap()
-    }
-}
-
-/// `DelayUs<u16>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayUs<u16> for Delay {
-    type Error = Infallible;
-
-    fn delay_us(&mut self, us: u16) -> Result<(), Self::Error> {
-        sleep(Duration::from_micros(us.into()));
-        Ok(())
+        DelayUs::delay_us(self, us as u32).unwrap()
     }
 }
 
 /// `DelayUs<u16>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayUs<u16> for Delay {
     fn delay_us(&mut self, us: u16) {
-        DelayUs::delay_us(self, us).unwrap()
+        DelayUs::delay_us(self, us as u32).unwrap()
     }
 }
 
-/// `DelayUs<u32>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayUs<u32> for Delay {
+/// `DelayUs` trait implementation for `embedded-hal` v1.0.0-alpha.7.
+impl DelayUs for Delay {
     type Error = Infallible;
 
     fn delay_us(&mut self, us: u32) -> Result<(), Self::Error> {
         sleep(Duration::from_micros(us.into()));
+        Ok(())
+    }
+
+    fn delay_ms(&mut self, ms: u32) -> Result<(), Self::Error> {
+        sleep(Duration::from_millis(u64::from(ms)));
         Ok(())
     }
 }
@@ -165,20 +114,15 @@ impl embedded_hal_0::blocking::delay::DelayUs<u32> for Delay {
     }
 }
 
-/// `DelayUs<u64>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl DelayUs<u64> for Delay {
-    type Error = Infallible;
-
-    fn delay_us(&mut self, us: u64) -> Result<(), Self::Error> {
-        sleep(Duration::from_micros(us));
-        Ok(())
-    }
-}
-
 /// `DelayUs<u64>` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::blocking::delay::DelayUs<u64> for Delay {
-    fn delay_us(&mut self, us: u64) {
-        DelayUs::delay_us(self, us).unwrap()
+    fn delay_us(&mut self, mut us: u64) {
+        while us > (u32::MAX as u64) {
+            us -= u32::MAX as u64;
+            DelayUs::delay_us(self, u32::MAX).unwrap();
+        }
+
+        DelayUs::delay_us(self, us as u32).unwrap()
     }
 }
 
@@ -220,31 +164,6 @@ impl Default for Timer {
     }
 }
 
-/// `CountDown` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl CountDown for Timer {
-    type Time = Duration;
-    type Error = Infallible;
-
-    /// Starts the timer with a `timeout`.
-    fn start<T>(&mut self, timeout: T) -> Result<(), Self::Error>
-    where
-        T: Into<Self::Time>,
-    {
-        self.start = Instant::now();
-        self.duration = timeout.into();
-        Ok(())
-    }
-
-    /// Returns `Ok` if the timer has wrapped.
-    fn wait(&mut self) -> nb::Result<(), Self::Error> {
-        if self.start.elapsed() >= self.duration {
-            Ok(())
-        } else {
-            Err(nb::Error::WouldBlock)
-        }
-    }
-}
-
 /// `CountDown` trait implementation for `embedded-hal` v0.2.6.
 impl embedded_hal_0::timer::CountDown for Timer {
     type Time = Duration;
@@ -254,7 +173,8 @@ impl embedded_hal_0::timer::CountDown for Timer {
     where
         T: Into<Self::Time>,
     {
-        CountDown::start(self, timeout).unwrap()
+        self.start = Instant::now();
+        self.duration = timeout.into();
     }
 
     /// Returns `Ok` if the timer has wrapped.