+++ src/uart/hal.rs
@@ -18,14 +18,22 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use embedded_hal::serial::nb::{Read, Write};
+use embedded_hal::serial::{self, ErrorType, nb::{Read, Write}};
 
 use super::{Error, Queue, Uart};
 
-/// `Read<u8>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
-impl Read<u8> for Uart {
+impl ErrorType for Uart {
     type Error = Error;
+}
 
+impl serial::Error for Error {
+  fn kind(&self) -> serial::ErrorKind {
+      serial::ErrorKind::Other
+  }
+}
+
+/// `Read<u8>` trait implementation for `embedded-hal` v1.0.0-alpha.7.
+impl Read<u8> for Uart {
     fn read(&mut self) -> nb::Result<u8, Self::Error> {
         let mut buffer = [0u8; 1];
         if Uart::read(self, &mut buffer)? == 0 {
@@ -45,10 +53,8 @@ impl embedded_hal_0::serial::Read<u8> for Uart {
     }
 }
 
-/// `Write<u8>` trait implementation for `embedded-hal` v1.0.0-alpha.5.
+/// `Write<u8>` trait implementation for `embedded-hal` v1.0.0-alpha.7.
 impl Write<u8> for Uart {
-    type Error = Error;
-
     fn write(&mut self, word: u8) -> nb::Result<(), Self::Error> {
         if Uart::write(self, &[word])? == 0 {
             Err(nb::Error::WouldBlock)